% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbouNationalPriors.R
\name{bbouNationalPriors}
\alias{bbouNationalPriors}
\title{Get informative priors for bboutools logistic models}
\usage{
bbouNationalPriors(
  Anthro,
  fire_excl_anthro,
  month = TRUE,
  force_run_model = FALSE
)
}
\arguments{
\item{Anthro}{numeric. Percent non-overlapping buffered anthropogenic
disturbance.}

\item{fire_excl_anthro}{numeric. Percent fire not overlapping with
anthropogenic disturbance.}

\item{month}{logical or "both". Should the returned survival be monthly (default)
or annual. If "both" then monthly survival with the name priors_survival_monthly
is returned as well as priors_survival.}

\item{force_run_model}{logical. Default is FALSE which means the priors are
drawn from the pre-calculated \code{bbou_national_priors_table}. When TRUE the
model is run to calculate the priors.}
}
\value{
a list with values:
\itemize{
\item priors_survival = c(b0_mu,b0_sd). Passed to \code{bboutools::bb_fit_survival}.
See \code{bb_priors_survival} for details.
\item priors_recruitment=c(b0_mu,b0_sd). Passed to \code{bboutools::bb_fit_recruitment}.
See \code{bb_priors_recruitment} for details.
}
}
\description{
Returns intercept prior parameter values for bboutools logistic models
that are informed by national demographic-disturbance relationships.
}
\details{
This is done by ... TO DO.

See \code{caribouMetrics::betaNationalPriors()} for additional details.
}
\examples{
nat_priors <- bbouNationalPriors(Anthro=50,fire=5)
print(nat_priors)
}
